+++ Realm/Tests/ObjectTests.m
@@ -209,6 +209,30 @@
     XCTAssertEqualObjects(ownerArrayDog.dog.dogName, @"PJ");
 }
 
+- (void)testInitFromDictionaryMissingPropertyKey {
+    CompanyObject *co = nil;
+    XCTAssertThrows([[CompanyObject alloc] initWithObject:@{}]);
+    XCTAssertNoThrow(co = [[CompanyObject alloc] initWithObject:@{@"name": @"a"}]);
+    XCTAssertEqualObjects(co.name, @"a");
+    XCTAssertEqual(co.employees.count, 0);
+
+    OwnerObject *oo = nil;
+    XCTAssertNoThrow(oo = [[OwnerObject alloc] initWithObject:@{@"name": @"a"}]);
+    XCTAssertEqualObjects(oo.name, @"a");
+    XCTAssertNil(oo.dog);
+}
+
+- (void)testInitFromDictionaryPropertyKey {
+    CompanyObject *co = nil;
+    XCTAssertNoThrow((co = [[CompanyObject alloc] initWithObject:@{@"name": @"a", @"employees": NSNull.null}]));
+    XCTAssertEqualObjects(co.name, @"a");
+    XCTAssertEqual(co.employees.count, 0);
+
+    OwnerObject *oo = nil;
+    XCTAssertNoThrow((oo = [[OwnerObject alloc] initWithObject:@{@"name": @"a", @"employees": NSNull.null}]));
+    XCTAssertEqualObjects(oo.name, @"a");
+    XCTAssertNil(oo.dog);
+}
 
 -(void)testObjectInitWithObjectTypeOther
 {