+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 ### Enhancements
 
 * Add support for querying for nil object properties.
+* Improve error message when specifying invalid literals when creating or 
+  initializing RLMObjects.
 
 ### Bugfixes