+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-0.83.0 Release notes (YYYY-MM-DD)
+0.83.0 Release notes (2014-08-13)
 =============================================================
 
 ### API breaking changes